A deamonized yara connector for cbr
