+++ MANIFEST
@@ -1,0 +1,4 @@
+/usr/share/cb/integrations/yaraconnector/yaraconnector
+%dir /usr/share/cb/integrations/yaraconnector
+%dir /var/log/cb/integrations/yaraconnector
+/etc/init/yaraconnector.conf
