+++ init-scripts/yaraconnector.conf
@@ -1,0 +1,13 @@
+description "Cb Response Yara Connector"
+author "zestep@vmware.com"
+
+start on (started network)
+stop on runlevel [!2345]
+
+respawn
+
+pre-start script
+    /usr/share/cb/integrations/yaraconnector/yaraconnector --lock-file /var/run/yaraconnector.pid --validate-rules  --config-file /etc/cb/integrations/yaraconnector/yaraconnector.conf &> /var/log/cb/integrations/yaraconnector/yaraconnector_config_check.log
+end script
+
+exec /usr/share/cb/integrations/yaraconnector/yaraconnector --lock-file /var/run/yaraconnector.pid --config-file /etc/cb/integrations/yaraconnector/yaraconnector.conf &> /var/log/cb/integrations/yaraconnector/yaraconnector.log
