+++ requirements.txt
@@ -1,37 +1,9 @@
-altgraph==0.16.1
-amqp==2.4.0
-attrdict==2.0.0
-billiard==3.5.0.5
-cachetools==3.0.0
-cbapi==1.3.6
-celery==4.2.1
-certifi==2018.11.29
-chardet==3.0.4
-future==0.17.1
-humanfriendly==4.17
-idna==2.8
-kombu==4.2.2.post1
-macholib==1.11
-pbr==5.1.1
-peewee==3.8.1
-pefile==2018.8.8
-pika==0.12.0
-prompt-toolkit==2.0.7
-protobuf==3.6.1
-psycopg2==2.7.6.1
-psycopg2-binary==2.7.6.1
-Pygments==2.3.1
-PyInstaller==3.4
-python-dateutil==2.7.5
-pytz==2018.9
-PyYAML==3.13
-redis==3.0.1
-requests==2.21.0
-six==1.12.0
-solrq==1.1.1
-tendo==0.2.12
-urllib3==1.24.1
-vine==1.2.0
-wcwidth==0.1.7
-yara-python==3.8.1
+celery
+redis
 humanfriendly
+lockfile
+peewee
+psycopg2-binary
+python-daemon
+requests
+yara-python
