+++ src/analysis_result.py
@@ -1,3 +1,6 @@
+import json
+from datetime import datetime
+
 class AnalysisResult(object):
     def __init__(self,
                  md5,
@@ -20,3 +23,11 @@ class AnalysisResult(object):
         self.stop_future_scans = stop_future_scans
         self.binary_not_available = binary_not_available
         self.misc = misc
+
+    def toJSON(self):
+        dict_rep = self.__dict__
+        for key, value in dict_rep.items():
+            if isinstance(value, datetime):
+                dict_rep[key] = value.timetuple()
+
+        return json.dumps(dict_rep)
