+++ src/exceptions.py
